config/homepage/custom/requirements.txt
flask
requests
os
python-dotenv
